backend/requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
chromedriver-autoinstaller
beautifulsoup4
python-dotenv
APIRouter
fitz
google-generativeai
motor
